src/requirements.txt
PySide6>=6.5.0
Pillow>=10.0.0
requests>=2.31.0
pytest>=7.4.0
pytest-qt>=4.2.0